Add validate function to Message class
Without this commit, validation of the data in a message object must be
done manually using a very specific `Rschema` validate method. While it
is not impossible to use this method directly, it is cumbersome to
repeat over and over. This method was being used by repeating the same
code in two places within the `Message` object.

This commit adds a new method to the `Message` object that can be easily run
to do the validation. The `Message` object has been updated internally to
utilize the new method. This method can also easily be used to
validate the object prior to sending it.