Add default message expiration if not set
Without this commit, the expriation of a message must be set explicitly
during construction of the message. Messages that are sent internally,
for things like association, use an expiration of 3 seconds. This
appears to be a reasonable default for most messages.

This commit ensures that a default expiration of 3 seconds will be set
on message send unless an expiration has been explictly set prior to
sending.